+++ common/src/main/java/com/alien/common/data/AlienAdvancements.java
@@ -19,6 +19,8 @@ public class AlienAdvancements {
 
     public static final BLibAdvancement DEFEAT_A_RAID = create("defeat_a_raid");
 
+    public static final BLibAdvancement DUAL_VARIANT_RAIDS = create("dual_variant_raids");
+
     public static final BLibAdvancement REMOVE_EMBRYO_WITH_CHORUS_FRUIT = create("remove_embryo_with_chorus_fruit");
 
     public static final BLibAdvancement ROOT = create("root");